gt58wifi: use stock media_profiles.xml for camera
Thanks to xda user named dylux for mentioning that he used
stock media_profiles.xml for camera.  This fixes the selfie camera
from crashing in video and picture mode.

However, it's not 100% perfect as the preview window stops
working when taking a photo (front and back camera).

To unfreeze, switch the camera front to back or back to front
and then preview will work